requirements.txt
protocolinterface
psutil
pyyaml
jinja2